try to add authProvider for postgRest via pgjwt
try to add authProvider for postgRest via pgjwt
