+++ package.json
@@ -1,13 +1,13 @@
 {
 	"name": "ra-postgrest-client",
-	"version": "0.1.2",
-	"description":
-		"dataProvider for [react-admin](https://marmelab.com/react-admin/) ( admin-on-rest 2.0 ) to access [postgrest](https://postgrest.com) REST service",
+	"version": "0.1.3",
+	"description": "dataProvider for [react-admin](https://marmelab.com/react-admin/) ( admin-on-rest 2.0 ) to access [postgrest](https://postgrest.com) REST service",
 	"main": "lib/index.js",
 	"repository": "https://github.com/tsingson/ra-postgrest-client",
 	"author": "tsingsonqin <tsingson@gmail.com>",
 	"license": "MIT",
 	"dependencies": {
+		"axios": "^0.18.0",
 		"ra-core": "^2.0.0-RC2",
 		"react-admin": "^2.0.0-RC2"
 	},
